move Polygon_with_holes_2 in the Polygon package
with the agreement from TAU
