+++ Boolean_set_operations_2/doc/Boolean_set_operations_2/PackageDescription.txt
@@ -37,7 +37,6 @@ containment predicates.
 - `GpsTraitsGeneralPolygon_2`
 - `GpsTraitsGeneralPolygonWithHoles_2`
 - `GeneralPolygon_2`
-- `GeneralPolygonWithHoles_2`
 - `ArrangementDirectionalXMonotoneTraits_2`
 - `GeneralPolygonSetTraits_2`
 - `GeneralPolygonSetDcel`
@@ -45,7 +44,6 @@ containment predicates.
 - `GeneralPolygonSetDcelHalfedge`
 
 ## Classes ##
-- `CGAL::Polygon_with_holes_2<Kernel,Container>`
 - `CGAL::Polygon_set_2<Kernel,Container,Dcel>`
 - `CGAL::General_polygon_set_2<Traits,Dcel>`
 - `CGAL::General_polygon_2<ArrTraits>`
@@ -64,8 +62,5 @@ containment predicates.
 - \link boolean_symmetric_difference `CGAL::symmetric_difference()` \endlink
 - \link boolean_oriented_side `CGAL::oriented_side()` \endlink
 - \link boolean_connect_holes `CGAL::connect_holes()` \endlink
-- `operator<<()` for `CGAL::Polygon_with_holes_2`, `CGAL::General_polygon_2` and `CGAL::General_polygon_with_holes_2`
-- `operator>>()` for `CGAL::Polygon_with_holes_2`, `CGAL::General_polygon_2` and `CGAL::General_polygon_with_holes_2`
-
 
 */
