+++ Polygon/include/CGAL/Polygon_with_holes_2.h
@@ -1,10 +1,14 @@
-// Copyright (c) 2005  Tel-Aviv University (Israel).
-// All rights reserved.
+// Copyright (c) 2005
+// Utrecht University (The Netherlands),
+// ETH Zurich (Switzerland),
+// INRIA Sophia-Antipolis (France),
+// Max-Planck-Institute Saarbruecken (Germany),
+// and Tel-Aviv University (Israel).  All rights reserved.
 //
-// This file is part of CGAL (www.cgal.org).
-// You can redistribute it and/or modify it under the terms of the GNU
-// General Public License as published by the Free Software Foundation,
-// either version 3 of the License, or (at your option) any later version.
+// This file is part of CGAL (www.cgal.org); you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public License as
+// published by the Free Software Foundation; either version 3 of the License,
+// or (at your option) any later version.
 //
 // Licensees holding a valid commercial license may use this file in
 // accordance with the commercial license agreement provided with the software.
@@ -14,7 +18,7 @@
 //
 // $URL$
 // $Id$
-// 
+//
 //
 // Author(s)     : Baruch Zukerman <baruchzu@post.tau.ac.il>
 
@@ -28,9 +32,21 @@
 
 namespace CGAL {
 
+/*!
+\ingroup PkgPolygon2
+
+The class `Polygon_with_holes_2` models the concept `GeneralPolygonWithHoles_2`.
+It represents a linear polygon with holes. It is parameterized with two
+types (`Kernel` and `Container`) that are used to instantiate
+the type `Polygon_2<Kernel,Container>`. The latter is used to
+represents the outer boundary and the boundary of the holes (if any exist).
+
+\cgalModels `GeneralPolygonWithHoles_2`
+
+*/
 template <class Kernel,
           class Containter = std::vector<typename Kernel::Point_2> >
-class Polygon_with_holes_2 : 
+class Polygon_with_holes_2 :
   public General_polygon_with_holes_2<CGAL::Polygon_2<Kernel, Containter> >
 {
 public:
@@ -41,17 +57,17 @@ public:
   typedef typename Base::Size                        Size;
 
   /*! Default constructor. */
-  Polygon_with_holes_2 () : 
+  Polygon_with_holes_2 () :
     Base()
   {}
 
   /*! Constructor from the base class. */
-  Polygon_with_holes_2 (const Base& base) : 
+  Polygon_with_holes_2 (const Base& base) :
     Base (base)
   {}
 
   /*! Constructor from a polygon. */
-  explicit Polygon_with_holes_2 (const Polygon_2& pgn_boundary) : 
+  explicit Polygon_with_holes_2 (const Polygon_2& pgn_boundary) :
     Base (pgn_boundary)
   {}
 
@@ -59,18 +75,36 @@ public:
   template <class HolesInputIterator>
   Polygon_with_holes_2 (const Polygon_2& pgn_boundary,
                         HolesInputIterator h_begin,
-                        HolesInputIterator h_end) : 
+                        HolesInputIterator h_end) :
     Base (pgn_boundary, h_begin, h_end)
   {}
 
   /*! Obtain the bounding box of the polygon with holes */
-  Bbox_2 bbox() const { return this->outer_boundary().bbox(); }  
+  Bbox_2 bbox() const { return this->outer_boundary().bbox(); }
 };
 
 //-----------------------------------------------------------------------//
 //                          operator<<
 //-----------------------------------------------------------------------//
 
+/*!
+This operator exports a polygon with holes to the output stream `out`.
+
+An ASCII and a binary format exist. The format can be selected with
+the \cgal modifiers for streams, `set_ascii_mode()` and `set_binary_mode()`
+respectively. The modifier `set_pretty_mode()` can be used to allow for (a
+few) structuring comments in the output. Otherwise, the output would
+be free of comments. The default for writing is ASCII without
+comments.
+
+The number of points of the outer boundary is exported followed by the
+points themselves in counterclockwise order. Then, the number of holes
+is exported, and for each hole, the number of points on its outer
+boundary is exported followed by the points themselves in clockwise
+order.
+
+\relates Polygon_with_holes_2
+*/
 template <class Kernel_, class Container_>
 std::ostream& operator<<(std::ostream &os,
                          const Polygon_with_holes_2<Kernel_, Container_>& p)
@@ -117,6 +151,19 @@ std::ostream& operator<<(std::ostream &os,
 //                          operator>>
 //-----------------------------------------------------------------------//
 
+/*!
+This operator imports a polygon with holes from the input stream `in`.
+
+An ASCII and a binary format exist. The stream detects the format
+automatically and can read both.
+
+The format consists of the number of points of the outer boundary followed
+by the points themselves in counterclockwise order, followed by the number of holes,
+and for each hole, the number of points of the outer boundary is followed
+by the points themselves in clockwise order.
+
+\relates Polygon_with_holes_2
+*/
 template <class Kernel_, class Container_>
 std::istream &operator>>(std::istream &is,
                          Polygon_with_holes_2<Kernel_, Container_>& p)
@@ -148,7 +195,7 @@ template <class Kernel_, class Container_>
 bool operator==(const Polygon_with_holes_2<Kernel_, Container_>& p1,
                 const Polygon_with_holes_2<Kernel_, Container_>& p2)
 {
-  typedef typename 
+  typedef typename
     Polygon_with_holes_2<Kernel_, Container_>::Hole_const_iterator HCI;
   typedef CGAL::Polygon_2<Kernel_, Container_> Polygon_2;
   if(&p1 == &p2)
